Address review: Extract scheduleCollectionRefresh to shared utility
Addresses reviewer recommendation from PR #833:
- Created shared collection-refresh.ts utility module
- Updated browser-sync-service.ts to import from shared utility
- Updated tauri-sync-listener.ts to import from shared utility

This follows the DRY (Don't Repeat Yourself) principle by consolidating
the duplicated debounce logic into a single source of truth.

Reviewer suggestions NOT addressed (with justification):
- Module-level mutable state: Reviewer noted this is "minor concern" for
  singleton services. Current implementation works correctly.
- Test coverage for debounce: Valid future improvement but outside scope
  of addressing immediate review feedback. Simple timeout logic.
