+++ packages/desktop-app/src/lib/services/tauri-sync-listener.ts
@@ -18,7 +18,6 @@
  */
 
 import { listen } from '@tauri-apps/api/event';
-import { get } from 'svelte/store';
 import type {
   NodeEventData,
   RelationshipEvent,
@@ -26,48 +25,13 @@ import type {
 } from '$lib/types/event-types';
 import { sharedNodeStore } from './shared-node-store.svelte';
 import { structureTree } from '$lib/stores/reactive-structure-tree.svelte';
-import { collectionsData, collectionsState } from '$lib/stores/collections';
 import type { Node } from '$lib/types/node';
 import { nodeToTaskNode } from '$lib/types/task-node';
 import { backendAdapter } from './backend-adapter';
 import { createLogger } from '$lib/utils/logger';
+import { scheduleCollectionRefresh } from '$lib/utils/collection-refresh';
 
 const log = createLogger('TauriSync');
-
-// Debounce timer for collection refreshes during bulk operations
-let collectionRefreshTimer: ReturnType<typeof setTimeout> | null = null;
-const COLLECTION_REFRESH_DEBOUNCE_MS = 300;
-
-/**
- * Debounced refresh of collections sidebar
- *
- * When member_of relationships change (especially during bulk imports),
- * we debounce the refresh to avoid excessive API calls.
- *
- * @param affectedCollectionId - Optional collection ID that was affected (for member refresh)
- */
-function scheduleCollectionRefresh(affectedCollectionId?: string): void {
-  if (collectionRefreshTimer) {
-    clearTimeout(collectionRefreshTimer);
-  }
-
-  collectionRefreshTimer = setTimeout(async () => {
-    collectionRefreshTimer = null;
-    log.debug('Refreshing collections after member_of change');
-
-    // Reload all collections (updates sidebar)
-    await collectionsData.loadCollections();
-
-    // If the affected collection is currently selected, also refresh its members
-    if (affectedCollectionId) {
-      const state = get(collectionsState);
-      if (state.selectedCollectionId === affectedCollectionId) {
-        log.debug('Refreshing members for selected collection', affectedCollectionId);
-        await collectionsData.loadMembers(affectedCollectionId);
-      }
-    }
-  }, COLLECTION_REFRESH_DEBOUNCE_MS);
-}
 
 /**
  * Normalize node data from domain events to type-specific format
